# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        '200':
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: status code
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Message'
    patch:
      summary: Patch message by id
      tags:
        - messages
      requestBody:
        description: id and patch 
        reqired: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: status code
          content:
            application/json:
             schema:
               $ref: '#/components/schemas/Message'
    delete:
      summary: delete message by id
      tags:
        - messages
      requestBody:
        description: id to del
        reqired: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '204':
          description: status code
          content:
            application/json:
             schema:
               $ref: '#/components/schemas/Message'
  /users:
    get:
      summary: Get all users
      tags: 
        - users
      responses:
        '200':
          description: a List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: new user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: create y/n?
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Ans'
    patch:
      summary: Change users data 
      tags:
        - users
      requestBody:
        description: data to change
        required: true 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: changed y/n?
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ans'
    delete:
      summary: Delete user by id
      tags:
        - users
      requestBody:
        dscription: data to change
        reqired: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: changed y/n?
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ans'
              
                  
              
        
              

      
              
components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        email: 
          type: string
        password:
          type: string
    Ans:
      type: object
      properties:
        id:
          type: integer
          format: uint
        Changed:
          type: boolean